# docs/swagger.yaml
basePath: /
definitions:
...
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  dto.AdminCheckResponse:
    properties:
      isAdmin:
        description: Whether have admin privileges // 是否具有管理员权限
        type: boolean
    type: object
  dto.AdminCloudflareConfig:
    properties:
      enabled:
        description: Whether to enable cloudflare tunnel // 是否启用 cloudflare 隧道
        type: boolean
      logEnabled:
        description: Whether to enable cloudflare tunnel logging // 是否开启 cloudflare
          隧道日志
        type: boolean
      token:
        description: cloudflare tunnel token // cloudflare 隧道令牌
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/haierkeys
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
  /api/admin/check:
    get:
      description: Check if the current logged-in user has system admin privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminCheckResponse'
              type: object
      security:
      - UserAuthToken: []
      summary: Check admin permission
      tags:
      - Config
  /api/admin/cloudflared_tunnel_download:
    get:
      description: Trigger the download of cloudflared binary for the current platform
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Download cloudflared binary
      tags:
      - System
  /api/admin/config:
    get:
      description: Get full system configuration information, requires admin privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get full admin config
      tags:
      - Config
    post:
      consumes:
      - application/json
      description: Modify full system configuration information, requires admin privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.AdminConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update admin config
      tags:
      - Config
  /api/admin/config/cloudflare:
    get:
      description: Get Cloudflare tunnel configuration, requires admin privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminCloudflareConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get Cloudflare config
      tags:
      - Config
    post:
      consumes:
      - application/json
      description: Modify Cloudflare tunnel configuration, requires admin privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCloudflareConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminCloudflareConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update Cloudflare config
      tags:
      - Config
  /api/admin/config/user_database:
    get:
      description: Get user database configuration information, requires admin privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminUserDatabaseConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get user database config
      tags:
      - Config
    post:
      consumes:
      - application/json
      description: Modify user database configuration information, requires admin
        privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUserDatabaseConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminUserDatabaseConfig'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update user database config
      tags:
      - Config
  /api/admin/config/user_database/test:
    post:
      consumes:
      - application/json
      description: Test if the provided database configuration can connect successfully,
        requires admin privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUserDatabaseConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
        "400":
          description: Connection failed
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Test user database connection
      tags:
      - Config
  /api/admin/gc:
    get:
      description: Manually run Go runtime GC and release memory to OS, requires admin
        privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Trigger manual GC
      tags:
      - System
  /api/admin/restart:
    get:
      description: Gracefully restart the server
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Trigger server restart
      tags:
      - System
...
  /api/admin/upgrade:
    get:
      description: Download latest version and restart server
      parameters:
      - description: Version to upgrade (e.g. 2.0.10 or latest)
        in: query
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Trigger server upgrade
      tags:
      - System
  /api/admin/users/create:
    post:
      consumes:
      - application/json
      description: Create a new user, requires admin privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDTO'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Create a new user
      tags:
      - Config
  /api/admin/users/list:
    get:
      description: Handle request to get all users.
      parameters:
      - description: Page number // 页码
        in: query
        name: page
        type: integer
      - description: Page size // 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.UserDTO'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get all users
      tags:
      - Config
  /api/admin/users/update:
    post:
      consumes:
      - application/json
      description: Update a user, requires admin privileges
      parameters:
      - description: Config Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserDTO'
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update a user
      tags:
      - Config
  /api/admin/ws_client/{traceId}:
    delete:
      description: Kick a WebSocket client by TraceID, requires admin privileges
      parameters:
      - description: Trace ID of the client
        in: path
        name: traceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Kick a WebSocket client
      tags:
      - System
  /api/admin/ws_clients:
    get:
      description: Get a list of all current WebSocket connections, requires admin
        privileges
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/app.WSClientInfo'
                  type: array
              type: object
        "403":
          description: Insufficient privileges
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get connected WebSocket clients
      tags:
      - System
  /api/auth/logout:
    post:
      description: |-
        Handle user logout HTTP request, revoke current auth token.
        处理用户退出登录 HTTP 请求，注销当前认证 Token。
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: User logout
      tags:
      - User
  /api/automations/executions:
    get:
      parameters:
      - in: query
        name: triggerId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.AutomationExecutionDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: List automation executions
      tags:
      - Automation
  /api/automations/executions/retry:
    post:
      consumes:
      - application/json
      parameters:
      - description: Execution ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AutomationExecutionRetryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Retry automation execution
      tags:
      - Automation
  /api/backup/config:
    delete:
      parameters:
      - description: ID // ID
        example: 1
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
        "400":
          description: Invalid Params
          schema:
            $ref: '#/definitions/app.Res'
        "401":
          description: Token Required
          schema:
            $ref: '#/definitions/app.Res'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Delete backup configuration
      tags:
      - Backup
    post:
      consumes:
      - application/json
      parameters:
      - description: Backup Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.BackupConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.BackupConfigDTO'
              type: object
        "400":
          description: Invalid Params
          schema:
            $ref: '#/definitions/app.Res'
        "401":
          description: Token Required
          schema:
            $ref: '#/definitions/app.Res'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Update backup configuration
      tags:
      - Backup
  /api/backup/configs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BackupConfigDTO'
                  type: array
              type: object
        "401":
          description: Token Required
          schema:
            $ref: '#/definitions/app.Res'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get backup configurations
      tags:
      - Backup
  /api/backup/historys:
    get:
      parameters:
      - description: Config ID // 配置 ID
        example: 1
        in: query
        name: configId
        required: true
        type: integer
      - description: Page number // 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: Page size // 每页大小
        example: 10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.BackupHistoryDTO'
                        type: array
                    type: object
              type: object
        "400":
          description: Invalid Params
          schema:
            $ref: '#/definitions/app.Res'
        "401":
          description: Token Required
          schema:
            $ref: '#/definitions/app.Res'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Get backup history list
      tags:
      - Backup
...
  /api/file/info:
    get:
      description: Get attachment metadata (FileDTO) by path
      parameters:
      - description: Context // 同步上下文
        example: ctx123
        in: query
        name: context
        type: string
      - description: Is in recycle bin // 是否在回收站
        example: false
        in: query
        name: isRecycle
        type: boolean
      - description: File path // 文件路径
        example: Image.png
        in: query
        name: path
        required: true
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Get attachment info
      tags:
      - File
  /api/file/recycle-clear:
    delete:
      consumes:
      - application/json
      description: Permanently clear selected files from recycle bin
      parameters:
      - description: Clear Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.FileRecycleClearRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Clear recycle bin
      tags:
      - File
  /api/file/rename:
    post:
      consumes:
      - application/json
      description: Rename an attachment to a new path
      parameters:
      - description: Rename Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.FileRenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Rename attachment
      tags:
      - File
  /api/file/restore:
    put:
      description: Restore deleted attachment from trash
      parameters:
      - description: Restore Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.FileRestoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FileDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Restore attachment
      tags:
      - File
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"